requirements.txt
flask==2.3.3
pandas==2.1.0
openai==1.3.0
python-dotenv==1.0.0 